+++ test/dummy/config/importmap.rb
@@ -6,10 +6,12 @@ pin "@hotwired/stimulus", to: "stimulus.min.js"
 pin "@hotwired/stimulus-loading", to: "stimulus-loading.js"
 pin_all_from "app/javascript/controllers", under: "controllers"
 
-# Pin FlatPack controllers
+# Pin FlatPack controllers. Nested TipTap is also imported from controllers/index.js
+# so TextArea rich_text hydrates on first paint.
 if defined?(FlatPack::Engine)
   pin_all_from FlatPack::Engine.root.join("app/javascript/flat_pack/controllers"), under: "controllers/flat_pack", to: "flat_pack/controllers", preload: false
   pin_all_from FlatPack::Engine.root.join("app/javascript/flat_pack/tiptap"), under: "flat_pack/tiptap", to: "flat_pack/tiptap", preload: false
+  pin "controllers/flat_pack/tiptap_controller", to: "flat_pack/controllers/tiptap_controller.js"
   pin "flat_pack/heroicons", to: "flat_pack/heroicons.js", preload: false
 end